Input processing migration from Camera to Player

+++ SnowEngine/SnowEngine/Vec3GLf.h
@@ -18,7 +18,7 @@ struct Vec3GLf {
 		return *this;
 	}
 
-	Vec3GLf& operator*=(double v) { // For scalar multiplication
+	Vec3GLf& operator*=(GLfloat v) { // For scalar multiplication
 		this->x *= v;
 		this->y *= v;
 		this->z *= v;
